Lots of reorganizations!
Spent a good amount of time reorganizing the current hierarchy.

Starting from table 2, I followed the naming convention that I created in table 1.

All standard well is named Well#. While all the lids of said wells are set up as children of each well and named to be Lid#.

Starting from table 3, we have smaller halved wells that I referred to as Tinywell#. Which is an empty that holds two tWell# that has childrens tLid# attach to them.

This will ensure maximum visibility and minimal cluttering. Please follow this naming convention if able.

In addition, there are also B1 which stands for "Bottle1" with c1 child for "caps1" and l1 child for "liquid1."

I have yet to get to changing all of them yet, since it will take some work to arrange them.

Also pending some improvements for the cloning system to make it move with bowl and align with the new tables.

Good progress today though.
Signing off.