Sorted out Table 3
Very exhausted today, couldn't done much. However, Table 3 is sorted now.

The amount of work makes me wonder if removing Table 4 then dup the 3 and make some changes would be more efficient.

That would involve some of the groups being messed up though, so I may just opt for the dumb way of powering through it, since this way the vertices are guaranteed to not be messed up.

Should have done the naming conventions on the get go. Luckily, it is still early enough to sort this out than having a cluster fuck later in the cycle.

Also some concerns with the well positions matching with cloning positions. I have a sneaky suspicion that this is not going to be as straight forward as it is appears to be.

We will see though, should be doable.