New Rail System Implemented
Implements a new rail system that controls both rotation and position using LerpRail.cs

The old systems and objects were moved to "Old" sections that can be relocated later if needed. The system can be understood from the comments in LerpRail.cs.

Includes a way to change position and rotation of camera when moving the rail. All new rail points must be added to the List of transforms in LerpRail.java that is in the MBP object.