Added randomization, mix score, order text
Randomization of each ingredients amount was added, will be modified to reflect real dishes later.
Score is now calculated when mixing to check if the ingredients are in the final recipe.
Order text now shows order with extras, but only for the first 10 seconds.
A block UI now appears for both arrows.